ココナラ用追加作成 関数を分けた lancers改良 driver_get()のパス回り改良、保存時のファイル処理改良

+++ functions.py
@@ -4,16 +4,23 @@ import pandas as pd
 import os
 import pathlib
 
+
 #webdriverの位置を都度変える
 def driver_get(url):
     driver_home = 'C:/chromedriver_win32/chromedriver'
     driver_gae = 'C:/Users/g2945/chromedriver/chromedriver'
     options = Options()
     #options.add_argument('--headless')
-    driver = webdriver.Chrome(driver_gae, options=options)
-    driver.implicitly_wait(10)
-    driver.get(url)
-    return driver
+    try:
+        driver = webdriver.Chrome(driver_home, options=options)
+        driver.implicitly_wait(10)
+        driver.get(url)
+        return driver
+    except:
+        driver2 = webdriver.Chrome(driver_gae, options=options)
+        driver2.implicitly_wait(10)
+        driver2.get(url)
+        return driver2
 
 
 def save_data(titles, urls, prices, name, kw):
@@ -29,10 +36,19 @@ def save_data(titles, urls, prices, name, kw):
     data_dir = os.path.join(cd, 'data')
     p_data_dir = os.path.join(parent_dir, 'data')
     dir_list = os.listdir(cd)
+
     if 'main.py' in dir_list:
-        df.to_csv(os.path.join(data_dir, '{}_{}.csv'.format(name, kw)), index=False)
+        if os.path.exists(data_dir):
+            df.to_csv(os.path.join(data_dir, '{}_{}.csv'.format(name, kw)), index=False)
+        else:
+            os.mkdir(data_dir)
+            df.to_csv(os.path.join(data_dir, '{}_{}.csv'.format(name, kw)), index=False)
     else:
-        df.to_csv(os.path.join(p_data_dir, '{}_{}.csv'.format(name, kw)), index=False)
+        if os.path.exists(p_data_dir):
+            df.to_csv(os.path.join(p_data_dir, '{}_{}.csv'.format(name, kw)), index=False)
+        else:
+            os.mkdir(p_data_dir)
+            df.to_csv(os.path.join(p_data_dir, '{}_{}.csv'.format(name, kw)), index=False)
 
 
 def print_data(name, kw):
